+++ package/example/lib/advanced_example/main.dart
@@ -1,4 +1,5 @@
 import 'package:easy_form_example/advanced_example/screens/fields_errors/fields_errors_screen.dart';
+import 'package:easy_form_example/advanced_example/screens/typed_result/typed_result_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:easy_form_example/advanced_example/screens/custom_field/custom_field_screen.dart';
 import 'package:easy_form_example/advanced_example/screens/login_demo/login_demo_screen.dart';
@@ -74,6 +75,11 @@ class StartScreen extends StatelessWidget {
                   _DemoButton(
                     text: 'Error displayer',
                     builder: (context) => ErrorDisplayerScreen(),
+                  ),
+                  const SizedBox(height: 16),
+                  _DemoButton(
+                    text: 'Typed form',
+                    builder: (context) => TypedResultDemoScreen(),
                   ),
                 ],
               ),