+++ package/lib/easy_form_kit.dart
@@ -1,6 +1,7 @@
 library easy_form_kit;
 
 export 'src/easy_form.dart';
+export 'src/easy_data_form.dart';
 export 'src/text_form_field.dart';
 export 'src/easy_form_field_error.dart';
 export 'src/custom_form_field.dart';