+++ package/lib/src/buttons/save_button.dart
@@ -180,6 +180,7 @@ class EasyFormSaveButton extends EasyFormActionButton {
       default:
         indicator = CircularProgressIndicator(
           strokeWidth: 2,
+          color: color,
         );
         break;
     }